requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
channels==4.0.0
daphne==4.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
djangorestframework-simplejwt==5.3.2
Pillow==10.0.0
